Initial handling for the new cordova-serve module in cordova-lib.
1. Adds cordova-serve as another 'repo' (which is really pointing to the cordova-lib repo).
2. Adds the ability to work with prefixed tags, so cordova-serve can maintain versions independent of cordova-lib.

Note that this introduces cordova-serve as another repo as far as COHO is concerned. This works ok for now, but ultimately if we introduce more modules in the cordova-lib repo we should probably add the concept of modules or packages to repos (so cordova-serve would not exist as a 'repo' in COHO, but as a module within the cordova-lib repo) - then. among other benefits, we won't need custom handling for the fact that cordova-lib and cordova-serve live within subdirectories of their repo.