CB-9361 Make verify-tags working nice with Windows command prompt.
The verify-tags command currently relies on users entering a Ctrl+D to trigger the end of input. However, this doesn't work in the Windows command prompt.

I've modified this to use Node's readline module - read the input line by line, then process when an empty line is encountered. This way, the tags can be pasted then simply press Enter. Note that on Unix-like systems, Ctrl+D will still work (and if you pipe in a file that doesn't have a blank line at the end, that will still work), since the EOF will trigger the 'close' event.